backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "secture-cortex-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "structlog>=24.4.0",
    "python-multipart>=0.0.18",
    "pgvector>=0.3.6",
    "httpx>=0.28.0",
    "email-validator>=2.1.0",
    "deepgram-sdk>=3.7.0",
    "langgraph>=1.1,<2.0",
    "langchain-core>=1.2,<2.0",
    "langchain-openai>=1.1,<2.0",
    "langgraph-checkpoint-postgres>=3.0,<4.0",
    "langgraph-supervisor>=0.0.30",
    "psycopg[binary]>=3.1.0",
    "pymupdf>=1.25.0",
    "python-docx>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "fpdf2>=2.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [".venv", "migrations/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
